refactor interaction handlers
